released v1.2.1 [MILESTONE]

+++ test/ForceTouchApplication/platforms/ios/ForceTouchApplication/Plugins/cordova-plugin-forcetouch/CDVForceTouch.m
@@ -1,6 +1,6 @@
 /*
  * Project: cordova-plugin-forcetouch
- * Version: 1.2.0
+ * Version: 1.2.1
  * File: CDVForceTouch.m
  * Author: Matteo Pisani
  * E-Mail: matteo.pisani.91@gmail.com
@@ -88,7 +88,7 @@
                 [TouchData setObject:@"0.0" forKey:@"force"];
             [TouchData setObject:[NSString stringWithFormat:@"%f",touch.maximumPossibleForce] forKey:@"maximumPossibleForce"];
             NSDictionary* TouchDataObject = [NSDictionary dictionaryWithDictionary:TouchData];
-            [ForceTouchPoints setObject:TouchDataObject forKey:[NSString stringWithFormat:@"%d",touchIndex++]];
+            [ForceTouchPoints setObject:TouchDataObject forKey:[NSString stringWithFormat:@"%lu",(unsigned long)touchIndex++]];
         }
     }
 }
